@@ -1,10 +1,15 @@
 {
-    "name": "grimzy/laravel-mysql-spatial",
+    "name": "limenet/laravel-mysql-spatial",
     "description": "MySQL spatial data types extension for Laravel.",
+    "homepage": "https://github.com/limenet/laravel-mysql-spatial",
     "scripts": {
-        "test": "phpunit -c phpunit.xml.dist",
-        "test:unit": "phpunit -c phpunit.xml.dist --testsuite unit",
-        "test:integration": "phpunit -c phpunit.xml.dist --testsuite integration"
+        "post-autoload-dump": "@php ./vendor/bin/testbench package:discover --ansi",
+        "analyse": "vendor/bin/phpstan analyse",
+        "test": "vendor/bin/phpunit",
+        "test-coverage": "vendor/bin/phpunit --coverage",
+        "format": "vendor/bin/pint",
+        "post-update-cmd": "@composer bump -D",
+        "rector": "./vendor/bin/rector process"
     },
     "type": "library",
     "license": "MIT",
@@ -12,42 +17,59 @@
         {
             "name": "Joseph Estefane",
             "email": "estefanejoe@gmail.com"
+        },
+        {
+            "name": "Linus Metzler",
+            "email": "hi@linusmetzler.me"
         }
     ],
     "require": {
-        "php": ">=7.3",
-        "ext-pdo": "*",
+        "php": "^8.2",
         "ext-json": "*",
-        "illuminate/database": "^8.0",
+        "ext-pdo": "*",
         "geo-io/wkb-parser": "^1.0",
+        "illuminate/contracts": "^11.0",
+        "illuminate/database": "^11.0",
         "jmikola/geojson": "^1.0"
     },
+    "conflict": {},
     "require-dev": {
-        "phpunit/phpunit": "~6.5",
-        "laravel/laravel": "^8.0",
-        "doctrine/dbal": "^2.5",
-        "laravel/browser-kit-testing": "^2.0",
-        "mockery/mockery": "^1.3"
+        "driftingly/rector-laravel": "^1.1",
+        "laravel/browser-kit-testing": "^7.2",
+        "laravel/laravel": "^11.0.6",
+        "laravel/pint": "^1.15.1",
+        "mockery/mockery": "^1.6.11",
+        "nunomaduro/collision": "^8.1.1",
+        "nunomaduro/larastan": "^2.9.4",
+        "orchestra/testbench": "^9.0.4",
+        "phpstan/extension-installer": "^1.3.1",
+        "phpstan/phpstan-deprecation-rules": "^1.1.4",
+        "phpstan/phpstan-phpunit": "^1.3.16",
+        "phpunit/phpunit": "^10.5.18"
     },
     "autoload": {
         "psr-4": {
-            "Grimzy\\LaravelMysqlSpatial\\": "src/"
+            "Limenet\\LaravelMysqlSpatial\\": "src"
         }
     },
-    "autoload-dev" : {
-        "classmap" : [
-            "tests/Unit",
-            "tests/Integration"
-        ]
+    "autoload-dev": {
+        "psr-4": {
+            "Limenet\\LaravelMysqlSpatial\\Tests\\": "tests"
+        }
+    },
+    "config": {
+        "sort-packages": true,
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        }
     },
     "extra": {
-        "branch-alias": {
-            "dev-master": "4.0.x-dev"
-        },
         "laravel": {
             "providers": [
-                "Grimzy\\LaravelMysqlSpatial\\SpatialServiceProvider"
+                "Limenet\\LaravelMysqlSpatial\\SpatialServiceProvider"
             ]
         }
-    }
+    },
+    "minimum-stability": "stable",
+    "prefer-stable": true
 }
