@@ -1,15 +1,17 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Types;
+namespace Limenet\LaravelMysqlSpatial\Types;
 
 use GeoIO\WKB\Parser\Parser;
+use GeoJson\Feature\Feature;
 use GeoJson\GeoJson;
-use Grimzy\LaravelMysqlSpatial\Exceptions\UnknownWKTTypeException;
 use Illuminate\Contracts\Support\Jsonable;
+use Limenet\LaravelMysqlSpatial\Exceptions\UnknownWKBException;
+use Limenet\LaravelMysqlSpatial\Exceptions\UnknownWKTTypeException;
 
-abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializable
+abstract class Geometry implements \JsonSerializable, GeometryInterface, Jsonable
 {
-    protected static $wkb_types = [
+    protected static array $wkb_types = [
         1 => Point::class,
         2 => LineString::class,
         3 => Polygon::class,
@@ -19,24 +21,21 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         7 => GeometryCollection::class,
     ];
 
-    protected $srid;
-
-    public function __construct($srid = 0)
+    public function __construct(protected int $srid = 0)
     {
-        $this->srid = (int) $srid;
     }
 
-    public function getSrid()
+    public function getSrid(): int
     {
         return $this->srid;
     }
 
-    public function setSrid($srid)
+    public function setSrid(int $srid): void
     {
-        $this->srid = (int) $srid;
+        $this->srid = $srid;
     }
 
-    public static function getWKTArgument($value)
+    public static function getWKTArgument(string $value): string
     {
         $left = strpos($value, '(');
         $right = strrpos($value, ')');
@@ -44,35 +43,39 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         return substr($value, $left + 1, $right - $left - 1);
     }
 
-    public static function getWKTClass($value)
+    /** @return class-string<Geometry> */
+    public static function getWKTClass(string $value): string
     {
         $left = strpos($value, '(');
+
+        if ($left === false) {
+            throw new UnknownWKTTypeException('Could not parse '.$value);
+        }
+
         $type = trim(substr($value, 0, $left));
 
-        switch (strtoupper($type)) {
-            case 'POINT':
-                return Point::class;
-            case 'LINESTRING':
-                return LineString::class;
-            case 'POLYGON':
-                return Polygon::class;
-            case 'MULTIPOINT':
-                return MultiPoint::class;
-            case 'MULTILINESTRING':
-                return MultiLineString::class;
-            case 'MULTIPOLYGON':
-                return MultiPolygon::class;
-            case 'GEOMETRYCOLLECTION':
-                return GeometryCollection::class;
-            default:
-                throw new UnknownWKTTypeException('Type was '.$type);
-        }
+        return match (strtoupper($type)) {
+            'POINT' => Point::class,
+            'LINESTRING' => LineString::class,
+            'POLYGON' => Polygon::class,
+            'MULTIPOINT' => MultiPoint::class,
+            'MULTILINESTRING' => MultiLineString::class,
+            'MULTIPOLYGON' => MultiPolygon::class,
+            'GEOMETRYCOLLECTION' => GeometryCollection::class,
+            default => throw new UnknownWKTTypeException('Type was '.$type)
+        };
     }
 
-    public static function fromWKB($wkb)
+    public static function fromWKB(string $wkb): Geometry
     {
         $srid = substr($wkb, 0, 4);
-        $srid = unpack('L', $srid)[1];
+        $unpacked = unpack('L', $srid);
+
+        if ($unpacked === false) {
+            throw new UnknownWKBException($wkb);
+        }
+
+        $srid = $unpacked[1];
 
         $wkb = substr($wkb, 4);
         $parser = new Parser(new Factory());
@@ -87,34 +90,35 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         return $parsed;
     }
 
-    public static function fromWKT($wkt, $srid = null)
+    public static function fromWKT(string $wkt, int $srid = 0): static
     {
         $wktArgument = static::getWKTArgument($wkt);
 
         return static::fromString($wktArgument, $srid);
     }
 
-    public static function fromJson($geoJson)
+    public static function fromJson(string|GeoJson $geoJson): self
     {
         if (is_string($geoJson)) {
-            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson));
+            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson, flags: JSON_THROW_ON_ERROR));
         }
 
         if ($geoJson->getType() === 'FeatureCollection') {
             return GeometryCollection::fromJson($geoJson);
         }
 
+        /** @var Feature $geoJson */
         if ($geoJson->getType() === 'Feature') {
             $geoJson = $geoJson->getGeometry();
         }
 
-        $type = '\Grimzy\LaravelMysqlSpatial\Types\\'.$geoJson->getType();
+        $type = '\Limenet\LaravelMysqlSpatial\Types\\'.$geoJson?->getType();
 
         return $type::fromJson($geoJson);
     }
 
     public function toJson($options = 0)
     {
-        return json_encode($this, $options);
+        return json_encode($this, $options | JSON_THROW_ON_ERROR);
     }
 }
