@@ -1,69 +1,72 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Types;
+namespace Limenet\LaravelMysqlSpatial\Types;
 
 use GeoJson\GeoJson;
 use GeoJson\Geometry\MultiLineString as GeoJsonMultiLineString;
-use Grimzy\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException;
+use Limenet\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException;
+use RuntimeException;
 
-class MultiLineString extends GeometryCollection
+/**
+ * @implements GeometryInterface<LineString>
+ *
+ * @extends GeometryCollection<GeometryInterface>
+ */
+class MultiLineString extends GeometryCollection implements \Stringable, GeometryInterface
 {
     /**
      * The minimum number of items required to create this collection.
-     *
-     * @var int
      */
-    protected $minimumCollectionItems = 1;
+    protected int $minimumCollectionItems = 1;
 
     /**
      * The class of the items in the collection.
-     *
-     * @var string
      */
-    protected $collectionItemType = LineString::class;
+    protected string $collectionItemType = LineString::class;
 
-    public function getLineStrings()
+    public function getLineStrings(): array
     {
         return $this->items;
     }
 
-    public function toWKT()
+    public function toWKT(): string
     {
         return sprintf('MULTILINESTRING(%s)', (string) $this);
     }
 
-    public static function fromString($wktArgument, $srid = 0)
+    public static function fromString(string $wktArgument, int $srid = 0): static
     {
         $str = preg_split('/\)\s*,\s*\(/', substr(trim($wktArgument), 1, -1));
-        $lineStrings = array_map(function ($data) {
-            return LineString::fromString($data);
-        }, $str);
+
+        if ($str === false) {
+            throw new RuntimeException();
+        }
+
+        $lineStrings = array_map(fn ($data) => LineString::fromString($data), $str);
 
         return new static($lineStrings, $srid);
     }
 
-    public function __toString()
+    public function __toString(): string
     {
-        return implode(',', array_map(function (LineString $lineString) {
-            return sprintf('(%s)', (string) $lineString);
-        }, $this->getLineStrings()));
+        return implode(',', array_map(fn (LineString $lineString) => sprintf('(%s)', (string) $lineString), $this->getLineStrings()));
     }
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         $this->validateItemType($value);
 
         parent::offsetSet($offset, $value);
     }
 
-    public static function fromJson($geoJson)
+    public static function fromJson(string|GeoJson $geoJson): self
     {
         if (is_string($geoJson)) {
-            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson));
+            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson, flags: JSON_THROW_ON_ERROR));
         }
 
-        if (!is_a($geoJson, GeoJsonMultiLineString::class)) {
-            throw new InvalidGeoJsonException('Expected '.GeoJsonMultiLineString::class.', got '.get_class($geoJson));
+        if (! $geoJson instanceof GeoJsonMultiLineString) {
+            throw new InvalidGeoJsonException(GeoJsonMultiLineString::class, $geoJson::class);
         }
 
         $set = [];
@@ -80,9 +83,8 @@ class MultiLineString extends GeometryCollection
 
     /**
      * Convert to GeoJson Point that is jsonable to GeoJSON.
-     *
-     * @return \GeoJson\Geometry\MultiLineString
      */
+    #[\ReturnTypeWillChange]
     public function jsonSerialize()
     {
         $lineStrings = [];
