@@ -1,55 +1,56 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Types;
+namespace Limenet\LaravelMysqlSpatial\Types;
 
 use GeoJson\GeoJson;
 use GeoJson\Geometry\MultiPolygon as GeoJsonMultiPolygon;
-use Grimzy\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException;
+use Limenet\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException;
+use RuntimeException;
 
-class MultiPolygon extends GeometryCollection
+/**
+ * @implements GeometryInterface<MultiPolygon>
+ *
+ * @extends GeometryCollection<Polygon>
+ */
+class MultiPolygon extends GeometryCollection implements \Stringable, GeometryInterface
 {
     /**
      * The minimum number of items required to create this collection.
-     *
-     * @var int
      */
-    protected $minimumCollectionItems = 1;
+    protected int $minimumCollectionItems = 1;
 
     /**
      * The class of the items in the collection.
-     *
-     * @var string
      */
-    protected $collectionItemType = Polygon::class;
+    protected string $collectionItemType = Polygon::class;
 
-    public function toWKT()
+    public function toWKT(): string
     {
         return sprintf('MULTIPOLYGON(%s)', (string) $this);
     }
 
-    public function __toString()
+    public function __toString(): string
     {
-        return implode(',', array_map(function (Polygon $polygon) {
-            return sprintf('(%s)', (string) $polygon);
-        }, $this->items));
+        return implode(',', array_map(fn (Polygon $polygon) => sprintf('(%s)', (string) $polygon), $this->items));
     }
 
-    public static function fromString($wktArgument, $srid = 0)
+    public static function fromString(string $wktArgument, int $srid = 0): static
     {
         $parts = preg_split('/(\)\s*\)\s*,\s*\(\s*\()/', $wktArgument, -1, PREG_SPLIT_DELIM_CAPTURE);
+
+        if ($parts === false) {
+            throw new RuntimeException();
+        }
+
         $polygons = static::assembleParts($parts);
 
-        return new static(array_map(function ($polygonString) {
-            return Polygon::fromString($polygonString);
-        }, $polygons), $srid);
+        return new static(array_map(fn ($polygonString) => Polygon::fromString($polygonString), $polygons), $srid);
     }
 
     /**
      * Get the polygons that make up this MultiPolygon.
-     *
-     * @return array|Polygon[]
      */
-    public function getPolygons()
+    public function getPolygons(): array
     {
         return $this->items;
     }
@@ -66,19 +67,15 @@ class MultiPolygon extends GeometryCollection
      * "((0 0,4 0,4 4,0 4,0 0),(1 1,2 1,2 2,1 2,1 1))",
      * "((-1 -1,-1 -2,-2 -2,-2 -1,-1 -1))",
      * "((-1 -1,-1 -2,-2 -2,-2 -1,-1 -1))"
-     *
-     * @param array $parts
-     *
-     * @return array
      */
-    protected static function assembleParts(array $parts)
+    protected static function assembleParts(array $parts): array
     {
         $polygons = [];
         $count = count($parts);
 
         for ($i = 0; $i < $count; $i++) {
             if ($i % 2 !== 0) {
-                list($end, $start) = explode(',', $parts[$i]);
+                [$end, $start] = explode(',', (string) $parts[$i]);
                 $polygons[$i - 1] .= $end;
                 $polygons[++$i] = $start.$parts[$i];
             } else {
@@ -89,21 +86,21 @@ class MultiPolygon extends GeometryCollection
         return $polygons;
     }
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         $this->validateItemType($value);
 
         parent::offsetSet($offset, $value);
     }
 
-    public static function fromJson($geoJson)
+    public static function fromJson(string|GeoJson $geoJson): self
     {
         if (is_string($geoJson)) {
-            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson));
+            $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson, flags: JSON_THROW_ON_ERROR));
         }
 
-        if (!is_a($geoJson, GeoJsonMultiPolygon::class)) {
-            throw new InvalidGeoJsonException('Expected '.GeoJsonMultiPolygon::class.', got '.get_class($geoJson));
+        if (! $geoJson instanceof GeoJsonMultiPolygon) {
+            throw new InvalidGeoJsonException(GeoJsonMultiPolygon::class, $geoJson::class);
         }
 
         $set = [];
@@ -124,9 +121,8 @@ class MultiPolygon extends GeometryCollection
 
     /**
      * Convert to GeoJson MultiPolygon that is jsonable to GeoJSON.
-     *
-     * @return \GeoJson\Geometry\MultiPolygon
      */
+    #[\ReturnTypeWillChange]
     public function jsonSerialize()
     {
         $polygons = [];
