@@ -1,20 +1,18 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Connectors;
+namespace Limenet\LaravelMysqlSpatial\Connectors;
 
-use Grimzy\LaravelMysqlSpatial\MysqlConnection;
 use Illuminate\Database\Connectors\ConnectionFactory as IlluminateConnectionFactory;
+use Limenet\LaravelMysqlSpatial\MysqlConnection;
 use PDO;
 
 class ConnectionFactory extends IlluminateConnectionFactory
 {
     /**
-     * @param string       $driver
-     * @param \Closure|PDO $connection
-     * @param string       $database
-     * @param string       $prefix
-     * @param array        $config
-     *
+     * @param  string  $driver
+     * @param  \Closure|PDO  $connection
+     * @param  string  $database
+     * @param  string  $prefix
      * @return \Illuminate\Database\ConnectionInterface
      */
     protected function createConnection($driver, $connection, $database, $prefix = '', array $config = [])
