@@ -1,13 +1,13 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Eloquent;
+namespace Limenet\LaravelMysqlSpatial\Eloquent;
 
-use Grimzy\LaravelMysqlSpatial\Exceptions\SpatialFieldsNotDefinedException;
-use Grimzy\LaravelMysqlSpatial\Exceptions\UnknownSpatialFunctionException;
-use Grimzy\LaravelMysqlSpatial\Exceptions\UnknownSpatialRelationFunction;
-use Grimzy\LaravelMysqlSpatial\Types\Geometry;
-use Grimzy\LaravelMysqlSpatial\Types\GeometryInterface;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
+use Limenet\LaravelMysqlSpatial\Exceptions\SpatialFieldsNotDefinedException;
+use Limenet\LaravelMysqlSpatial\Exceptions\UnknownSpatialFunctionException;
+use Limenet\LaravelMysqlSpatial\Exceptions\UnknownSpatialRelationFunction;
+use Limenet\LaravelMysqlSpatial\Types\Geometry;
+use Limenet\LaravelMysqlSpatial\Types\GeometryInterface;
 
 /**
  * Trait SpatialTrait.
@@ -61,9 +61,8 @@ trait SpatialTrait
     /**
      * Create a new Eloquent query builder for the model.
      *
-     * @param \Illuminate\Database\Query\Builder $query
-     *
-     * @return \Grimzy\LaravelMysqlSpatial\Eloquent\Builder
+     * @param  \Illuminate\Database\Query\Builder  $query
+     * @return \Limenet\LaravelMysqlSpatial\Eloquent\Builder
      */
     public function newEloquentBuilder($query)
     {
@@ -117,13 +116,13 @@ trait SpatialTrait
         if (property_exists($this, 'spatialFields')) {
             return $this->spatialFields;
         } else {
-            throw new SpatialFieldsNotDefinedException(__CLASS__.' has to define $spatialFields');
+            throw new SpatialFieldsNotDefinedException(self::class.' has to define $spatialFields');
         }
     }
 
-    public function isColumnAllowed($geometryColumn)
+    public function isColumnAllowed($geometryColumn): bool
     {
-        if (!in_array($geometryColumn, $this->getSpatialFields())) {
+        if (! in_array($geometryColumn, $this->getSpatialFields())) {
             throw new SpatialFieldsNotDefinedException();
         }
 
@@ -163,7 +162,7 @@ trait SpatialTrait
 
         $columns = $query->getQuery()->columns;
 
-        if (!$columns) {
+        if (! $columns) {
             $query->select('*');
         }
 
@@ -206,7 +205,7 @@ trait SpatialTrait
 
         $columns = $query->getQuery()->columns;
 
-        if (!$columns) {
+        if (! $columns) {
             $query->select('*');
         }
         $query->selectRaw("st_distance_sphere(`$geometryColumn`, ST_GeomFromText(?, ?, 'axis-order=long-lat')) as distance", [
@@ -219,7 +218,7 @@ trait SpatialTrait
     {
         $this->isColumnAllowed($geometryColumn);
 
-        if (!in_array($relationship, $this->stRelations)) {
+        if (! in_array($relationship, $this->stRelations)) {
             throw new UnknownSpatialRelationFunction($relationship);
         }
 
@@ -275,7 +274,7 @@ trait SpatialTrait
     {
         $this->isColumnAllowed($geometryColumn);
 
-        if (!in_array($orderFunction, $this->stOrderFunctions)) {
+        if (! in_array($orderFunction, $this->stOrderFunctions)) {
             throw new UnknownSpatialFunctionException($orderFunction);
         }
 
