@@ -1,6 +1,6 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Schema;
+namespace Limenet\LaravelMysqlSpatial\Schema;
 
 use Illuminate\Database\Schema\Blueprint as IlluminateBlueprint;
 
@@ -8,129 +8,91 @@ class Blueprint extends IlluminateBlueprint
 {
     /**
      * Add a geometry column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function geometry($column, $srid = null)
+    public function geometry($column, $subtype = null, $srid = 0): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('geometry', $column, compact('srid'));
+        return $this->addColumn('geometry', $column, ['subtype' => $subtype, 'srid' => $srid]);
     }
 
     /**
      * Add a point column on the table.
      *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
+     * @param  ?int  $srid
      */
-    public function point($column, $srid = null)
+    public function point($column, $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('point', $column, compact('srid'));
+        return $this->addColumn('point', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a linestring column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function lineString($column, $srid = null)
+    public function lineString($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('linestring', $column, compact('srid'));
+        return $this->addColumn('linestring', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a polygon column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function polygon($column, $srid = null)
+    public function polygon($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('polygon', $column, compact('srid'));
+        return $this->addColumn('polygon', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a multipoint column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function multiPoint($column, $srid = null)
+    public function multiPoint($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('multipoint', $column, compact('srid'));
+        return $this->addColumn('multipoint', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a multilinestring column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function multiLineString($column, $srid = null)
+    public function multiLineString($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('multilinestring', $column, compact('srid'));
+        return $this->addColumn('multilinestring', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a multipolygon column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function multiPolygon($column, $srid = null)
+    public function multiPolygon($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('multipolygon', $column, compact('srid'));
+        return $this->addColumn('multipolygon', $column, ['srid' => $srid]);
     }
 
     /**
      * Add a geometrycollection column on the table.
-     *
-     * @param string   $column
-     * @param null|int $srid
-     *
-     * @return \Illuminate\Support\Fluent
      */
-    public function geometryCollection($column, $srid = null)
+    public function geometryCollection($column, ?int $srid = null): \Illuminate\Support\Fluent
     {
-        return $this->addColumn('geometrycollection', $column, compact('srid'));
+        return $this->addColumn('geometrycollection', $column, ['srid' => $srid]);
     }
 
     /**
      * Specify a spatial index for the table.
      *
-     * @param string|array $columns
-     * @param string       $name
-     *
-     * @return \Illuminate\Support\Fluent
+     * @param  string|array  $columns
+     * @param  string  $name
      */
-    public function spatialIndex($columns, $name = null)
+    public function spatialIndex($columns, $name = null): \Illuminate\Support\Fluent
     {
-        return $this->indexCommand('spatial', $columns, $name);
+        return $this->indexCommand(
+            'spatial',
+            $columns,
+            $name ?: $this->createIndexName('spatial', is_array($columns) ? $columns : [$columns])
+        );
     }
 
     /**
      * Indicate that the given index should be dropped.
      *
-     * @param string|array $index
-     *
-     * @return \Illuminate\Support\Fluent
+     * @param  string|array  $index
      */
-    public function dropSpatialIndex($index)
+    public function dropSpatialIndex($index): \Illuminate\Support\Fluent
     {
         return $this->dropIndexCommand('dropIndex', 'spatial', $index);
     }
