@@ -1,6 +1,6 @@
 <?php
 
-namespace Grimzy\LaravelMysqlSpatial\Schema;
+namespace Limenet\LaravelMysqlSpatial\Schema;
 
 use Closure;
 use Illuminate\Database\Schema\MySqlBuilder;
@@ -10,12 +10,10 @@ class Builder extends MySqlBuilder
     /**
      * Create a new command set with a Closure.
      *
-     * @param string  $table
-     * @param Closure $callback
-     *
+     * @param  string  $table
      * @return Blueprint
      */
-    protected function createBlueprint($table, Closure $callback = null)
+    protected function createBlueprint($table, ?Closure $callback = null)
     {
         return new Blueprint($table, $callback);
     }
