@@ -1,15 +1,15 @@
 <?php
 
-use Grimzy\LaravelMysqlSpatial\Eloquent\SpatialTrait;
+namespace Limenet\LaravelMysqlSpatial\Tests\Integration\Models;
+
 use Illuminate\Database\Eloquent\Model;
+use Limenet\LaravelMysqlSpatial\Eloquent\SpatialTrait;
 
 /**
- * Class WithSridModel.
- *
  * @property int                                          id
- * @property \Grimzy\LaravelMysqlSpatial\Types\Point      location
- * @property \Grimzy\LaravelMysqlSpatial\Types\LineString line
- * @property \Grimzy\LaravelMysqlSpatial\Types\LineString shape
+ * @property \Limenet\LaravelMysqlSpatial\Types\Point      location
+ * @property \Limenet\LaravelMysqlSpatial\Types\LineString line
+ * @property \Limenet\LaravelMysqlSpatial\Types\LineString shape
  */
 class WithSridModel extends Model
 {
