@@ -1,8 +1,12 @@
 <?php
 
-use Grimzy\LaravelMysqlSpatial\Types\LineString;
-use Grimzy\LaravelMysqlSpatial\Types\MultiLineString;
-use Grimzy\LaravelMysqlSpatial\Types\Point;
+namespace Limenet\LaravelMysqlSpatial\Tests\Unit\Types;
+
+use InvalidArgumentException;
+use Limenet\LaravelMysqlSpatial\Tests\Unit\BaseTestCase;
+use Limenet\LaravelMysqlSpatial\Types\LineString;
+use Limenet\LaravelMysqlSpatial\Types\MultiLineString;
+use Limenet\LaravelMysqlSpatial\Types\Point;
 
 class MultiLineStringTest extends BaseTestCase
 {
@@ -46,8 +50,8 @@ class MultiLineStringTest extends BaseTestCase
     public function testInvalidGeoJsonException()
     {
         $this->assertException(
-            \Grimzy\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException::class,
-            sprintf('Expected %s, got %s', GeoJson\Geometry\MultiLineString::class, GeoJson\Geometry\Point::class)
+            \Limenet\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException::class,
+            sprintf('Expected %s, got %s', \GeoJson\Geometry\MultiLineString::class, \GeoJson\Geometry\Point::class)
         );
         MultiLineString::fromJson('{"type":"Point","coordinates":[3.4,1.2]}');
     }
@@ -64,7 +68,7 @@ class MultiLineStringTest extends BaseTestCase
     {
         $this->assertException(
             InvalidArgumentException::class,
-            'Grimzy\LaravelMysqlSpatial\Types\MultiLineString must contain at least 1 entry'
+            'Limenet\LaravelMysqlSpatial\Types\MultiLineString must contain at least 1 entry'
         );
         $multilinestring = new MultiLineString([]);
     }
@@ -73,7 +77,7 @@ class MultiLineStringTest extends BaseTestCase
     {
         $this->assertException(
             InvalidArgumentException::class,
-            'Grimzy\LaravelMysqlSpatial\Types\MultiLineString must be a collection of Grimzy\LaravelMysqlSpatial\Types\LineString'
+            'Limenet\LaravelMysqlSpatial\Types\MultiLineString must be a collection of Limenet\LaravelMysqlSpatial\Types\LineString'
         );
         $multilinestring = new MultiLineString([
             new LineString([new Point(0, 0), new Point(1, 1)]),
@@ -109,7 +113,7 @@ class MultiLineStringTest extends BaseTestCase
         // assert invalid
         $this->assertException(
             InvalidArgumentException::class,
-            'Grimzy\LaravelMysqlSpatial\Types\MultiLineString must be a collection of Grimzy\LaravelMysqlSpatial\Types\LineString'
+            'Limenet\LaravelMysqlSpatial\Types\MultiLineString must be a collection of Limenet\LaravelMysqlSpatial\Types\LineString'
         );
         $multilinestring[] = 1;
     }
